docs: Change the docs color scheme to match the new logo better
This is totally unnecessary, but the current indigo color scheme doesn't really
match the blue in the new blue jay logo.

Setting totally custom colors in mkdocs is a pain since you need to override
lots of different css properties, but just using the built in 'blue' color
works pretty well. The 'light blue' color actually matches the logo better, but
I was a little concerned about readability for hyperlinks.

The accent color is overridden to 'blue' as well to match better. Note that
even though the primary and accent have the same name, they are actually
slightly different colors due to the way the material theme works.