cli_utils: command callers should now print snapshot stats
This means we can chose not to print the snapshot stats if we know we
will perform another snapshot operation during the same command, to
reduce the number of warnings and hints shown to the user.

Previously, print_snapshot_stats would also get called by maybe_snapshot
automatically (by calling maybe_snapshot_impl, where the auto print was
located). This is still the case.