cli: git push: rename `git.sign-on-push` to `git.sign-on-push.enabled` and add `git.sign-on-push.behavior`
Previously, when `git.sign-on-push` is set to `true`, commits would be
signed upon pushing with Git with the sign behavior of
`SignBehavior::Own`. However, this could not be customized like
`signing.sign-behavior`.

This commit renames `git.sign-on-push` to `git.sign-on-push.enabled` and
adds `git.sign-on-push.behavior` to allow users to customize the sign
behavior when pushing commits.